movie-webapp/src/app/header/header.component.ts
import { Component ,Input} from '@angular/core';
import { AuthService } from '../auth.service';
import { Router } from '@angular/router';
@Component({
  selector: 'app-header',
  templateUrl: './header.component.html',
  styleUrls: ['./header.component.css']
})
export class HeaderComponent {
  isLoggedIn: boolean = false;

  constructor(private authService: AuthService) {
    this.authService.isLoggedIn$.subscribe(loggedIn => {
      this.isLoggedIn = loggedIn;
    });
  }

  // Optionally, add a method to handle logout and update isLoggedIn
  logout() {
    this.authService.logout();
    this.isLoggedIn = false;
  }

}